Add files via upload
The databases as uploaded on a localhost server on XAMPP in MyPHPAdmin, including all the tables, records and field pertaining to the recipes, ingredients and the particular ingredients for each recipe are detailed here. The same are used as part of a prototype database, wherein which a certain number of ingredients and recipes are mentioned.